- NEW added CSS to browser view for Control and IItem objects - NEW added delete dialog to control view
git-svn-id: svn+ssh://svn.verinice.org/data/subversion/svnroot/TRUNK@1024 4c0145ca-515c-4d1a-a6bc-f9bd594c3410

+++ sernet.gs.ui.rcp.main/src/sernet/verinice/iso27k/rcp/CatalogView.java
@@ -30,6 +30,8 @@ import javax.swing.text.DateFormatter;
 import org.apache.log4j.Logger;
 import org.eclipse.jface.action.Action;
 import org.eclipse.jface.action.IToolBarManager;
+import org.eclipse.jface.dialogs.InputDialog;
+import org.eclipse.jface.dialogs.MessageDialog;
 import org.eclipse.jface.viewers.IStructuredContentProvider;
 import org.eclipse.jface.viewers.ITreeContentProvider;
 import org.eclipse.jface.viewers.LabelProvider;
@@ -264,7 +266,9 @@ public class CatalogView extends ViewPart {
 		
 		deleteCatalogAction = new Action() {
 			public void run() {
-				deleteCatalog();
+				boolean confirm = MessageDialog.openConfirm(viewer.getControl().getShell(), "Wirklich löschen?", "Diesen Katalog wirklich löschen?");
+				if (confirm)
+					deleteCatalog();
 			}
 		};
 		deleteCatalogAction.setText("Katalog löschen...");
